+++ src/api.rs
@@ -2,6 +2,7 @@ use crate::ApiStationClickResult;
 use crate::ApiStationHistory;
 use crate::ApiStationVoteResult;
 use crate::ApiStatus;
+use crate::RbError;
 use crate::external::post_api;
 use crate::ApiConfig;
 use crate::CountrySearchBuilder;
@@ -11,7 +12,6 @@ use crate::TagSearchBuilder;
 
 use serde::de::DeserializeOwned;
 use std::collections::HashMap;
-use std::error::Error;
 
 use rand::seq::SliceRandom;
 use rand::thread_rng;
@@ -27,10 +27,10 @@ use async_std_resolver::{config, resolver};
 /// 
 /// Example
 /// ```rust
-/// use std::error::Error;
+/// use radiobrowser::RbError;
 /// use radiobrowser::RadioBrowserAPI;
 /// #[async_std::main]
-/// async fn main() -> Result<(), Box<dyn Error>> {
+/// async fn main() -> Result<(), RbError> {
 ///     let mut api = RadioBrowserAPI::new().await?;
 ///     Ok(())
 /// }
@@ -45,7 +45,7 @@ impl RadioBrowserAPI {
     /// Create a new instance of a radiobrowser api client.
     /// It will fetch a list of radiobrowser server with get_default_servers()
     /// and save it internally.
-    pub async fn new() -> Result<Self, Box<dyn Error>> {
+    pub async fn new() -> Result<Self, RbError> {
         Ok(RadioBrowserAPI {
             servers: RadioBrowserAPI::get_default_servers().await?,
             current: 0,
@@ -54,7 +54,7 @@ impl RadioBrowserAPI {
 
     /// Create a new instance of a radiobrowser api client from
     /// a single dns name. Use this is you want to connect to a single named server.
-    pub async fn new_from_dns_a<P: AsRef<str>>(dnsname: P) -> Result<Self, Box<dyn Error>> {
+    pub async fn new_from_dns_a<P: AsRef<str>>(dnsname: P) -> Result<Self, RbError> {
         Ok(RadioBrowserAPI {
             servers: vec![dnsname.as_ref().to_string()],
             current: 0,
@@ -63,7 +63,7 @@ impl RadioBrowserAPI {
 
     /// Create a new instance of a radiobrowser api client from
     /// a dns srv record which may have multiple dns A/AAAA records.
-    pub async fn new_from_dns_srv<P: AsRef<str>>(srvname: P) -> Result<Self, Box<dyn Error>> {
+    pub async fn new_from_dns_srv<P: AsRef<str>>(srvname: P) -> Result<Self, RbError> {
         Ok(RadioBrowserAPI {
             servers: RadioBrowserAPI::get_servers_from_dns_srv(srvname).await?,
             current: 0,
@@ -83,12 +83,12 @@ impl RadioBrowserAPI {
     async fn post_api<P: DeserializeOwned, A: AsRef<str>>(
         &mut self,
         endpoint: A,
-    ) -> Result<P, Box<dyn Error>> {
+    ) -> Result<P, RbError> {
         let mapjson = HashMap::new();
         post_api(self.get_current_server(), endpoint.as_ref(), mapjson).await
     }
 
-    pub async fn get_station_changes(&mut self, limit: u64, last_change_uuid: Option<String>) -> Result<Vec<ApiStationHistory>, Box<dyn Error>> {
+    pub async fn get_station_changes(&mut self, limit: u64, last_change_uuid: Option<String>) -> Result<Vec<ApiStationHistory>, RbError> {
         let query = match last_change_uuid {
             Some(uuid) => format!("/json/stations/changed?limit={}&lastchangeuuid={}", limit, uuid),
             None => format!("/json/stations/changed?limit={}", limit)
@@ -96,21 +96,21 @@ impl RadioBrowserAPI {
         Ok(self.post_api(query).await?)
     }
 
-    pub async fn get_server_config(&mut self) -> Result<ApiConfig, Box<dyn Error>> {
+    pub async fn get_server_config(&mut self) -> Result<ApiConfig, RbError> {
         Ok(self.post_api("/json/config").await?)
     }
 
-    pub async fn get_server_status(&mut self) -> Result<ApiStatus, Box<dyn Error>> {
+    pub async fn get_server_status(&mut self) -> Result<ApiStatus, RbError> {
         Ok(self.post_api("/json/stats").await?)
     }
 
     /// Add a click to a station found by stationuuid
-    pub async fn station_click<P: AsRef<str>>(&mut self, stationuuid: P) -> Result<ApiStationClickResult, Box<dyn Error>> {
+    pub async fn station_click<P: AsRef<str>>(&mut self, stationuuid: P) -> Result<ApiStationClickResult, RbError> {
         Ok(self.post_api(format!("/json/url/{}",stationuuid.as_ref())).await?)
     }
 
     /// Add a vote to a station found by a stationuuid
-    pub async fn station_vote<P: AsRef<str>>(&mut self, stationuuid: P) -> Result<ApiStationVoteResult, Box<dyn Error>> {
+    pub async fn station_vote<P: AsRef<str>>(&mut self, stationuuid: P) -> Result<ApiStationVoteResult, RbError> {
         Ok(self.post_api(format!("/json/vote/{}",stationuuid.as_ref())).await?)
     }
 
@@ -134,16 +134,16 @@ impl RadioBrowserAPI {
         &mut self,
         endpoint: P,
         mapjson: HashMap<String, String>,
-    ) -> Result<Q, Box<dyn Error>> {
+    ) -> Result<Q, RbError> {
         post_api(self.get_current_server(), endpoint, mapjson).await
     }
 
-    pub async fn get_default_servers() -> Result<Vec<String>, Box<dyn Error>> {
+    pub async fn get_default_servers() -> Result<Vec<String>, RbError> {
         trace!("get_default_servers()");
         RadioBrowserAPI::get_servers_from_dns_srv("_api._tcp.radio-browser.info").await
     }
 
-    async fn get_servers_from_dns_srv<P: AsRef<str>>(srvname: P) -> Result<Vec<String>, Box<dyn Error>> {
+    async fn get_servers_from_dns_srv<P: AsRef<str>>(srvname: P) -> Result<Vec<String>, RbError> {
         trace!("get_servers_from_dns_srv()");
         let resolver = resolver(
             config::ResolverConfig::default(),